added login and signup compoenents and Views
